# docs/openapi.json
{
  "components" : {
    "responses" : {
      "BadRequest" : {
        "content" : {
          "application/problem+json" : {
            "examples" : {
              "validation-error" : {
                "summary" : "Invalid request",
                "value" : {
                  "errors" : [ {
                    "detail" : "must not be blank",
                    "pointer" : "#/title"
                  } ],
                  "status" : 400,
                  "title" : "Invalid request",
                  "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#validation-error"
                }
              }
            },
            "schema" : {
              "$ref" : "#/components/schemas/ValidationProblem"
            }
          }
        },
        "description" : "Invalid request: see errors for each invalid value."
      },
...
      "InvalidCredentials" : {
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Problem"
            }
          }
        },
        "description" : "Wrong email or password, or an unknown, used or revoked refresh token."
      },
...
      "TooManyRequests" : {
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Problem"
            }
          }
        },
        "description" : "Too many requests from this address or for this email.",
        "headers" : {
          "Retry-After" : {
            "description" : "Seconds to wait before trying again.",
            "schema" : {
              "type" : "integer"
            }
          }
        }
      },
      "Unauthorized" : {
        "description" : "Missing, expired or invalid access token.",
        "headers" : {
          "WWW-Authenticate" : {
            "description" : "Bearer, with the reason of the refusal.",
            "schema" : {
              "type" : "string"
            }
          }
        }
      },
      "UnexpectedError" : {
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Problem"
            }
          }
        },
        "description" : "An unexpected failure on the server."
      },
...
    },
    "schemas" : {
...
      "AuthResponseDto" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "refreshTokenExpiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "tokenType" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "InvalidValue" : {
        "description" : "One invalid value of a request.",
        "properties" : {
          "detail" : {
            "description" : "What is wrong with the value.",
            "type" : "string"
          },
          "parameter" : {
            "description" : "Name of the query, path, form or multipart parameter, when not in a JSON body.",
            "type" : "string"
          },
          "pointer" : {
            "description" : "JSON Pointer to the value in the JSON body, such as #/email; # is the body.",
            "type" : "string"
          }
        },
        "required" : [ "detail" ],
        "type" : "object"
      },
      "LoginRequestDto" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "minLength" : 1,
            "type" : "string"
          },
          "password" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "email", "password" ],
        "type" : "object"
      },
...
      "PasswordResetConfirmDto" : {
        "properties" : {
          "newPassword" : {
            "maxLength" : 128,
            "minLength" : 8,
            "type" : "string"
          },
          "token" : {
            "maxLength" : 128,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "newPassword", "token" ],
        "type" : "object"
      },
      "PasswordResetRequestDto" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "maxLength" : 320,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "email" ],
        "type" : "object"
      },
      "Problem" : {
        "description" : "A problem document (RFC 9457). docs/problems.md lists every problem and what to do.",
        "properties" : {
          "detail" : {
            "description" : "This occurrence, for a person to read. It follows Accept-Language: do not parse it.",
            "type" : "string"
          },
          "instance" : {
            "description" : "The path of the request.",
            "format" : "uri-reference",
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "maximum" : 599,
            "minimum" : 100,
            "type" : "integer"
          },
          "title" : {
            "description" : "Short summary, the same for every occurrence: the problem type's title, or the status phrase.",
            "type" : "string"
          },
          "type" : {
            "description" : "What went wrong; absent when the status says it all (about:blank). It links to the documentation of the problem.",
            "format" : "uri",
            "type" : "string"
          }
        },
        "required" : [ "status", "title" ],
        "type" : "object"
      },
      "RefreshTokenRequestDto" : {
        "properties" : {
          "refreshToken" : {
            "maxLength" : 128,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "refreshToken" ],
        "type" : "object"
      },
...
      "ValidationProblem" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Problem"
        }, {
          "properties" : {
            "errors" : {
              "description" : "Each invalid value, with the validation-error type.",
              "items" : {
                "$ref" : "#/components/schemas/InvalidValue"
              },
              "type" : "array"
            }
          },
          "type" : "object"
        } ]
      },
      "VerifyEmailRequestDto" : {
        "properties" : {
          "token" : {
            "maxLength" : 128,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "token" ],
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "description" : "The access token returned by POST /api/v1/auth/login or /refresh.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
...
  "info" : {
    "contact" : {
      "name" : "Tasks API on GitHub",
      "url" : "https://github.com/jubasse/spring-boot-tasks-api"
    },
    "description" : "Tasks, with their comments, files and history, for teams whose members sign up themselves.\n\n**Authentication.** Sign in with `POST /api/v1/auth/login`, then send the access token in an `Authorization: Bearer` header. It lasts 15 minutes: `POST /api/v1/auth/refresh` trades the refresh token for a new pair. Task operations also need an active account: a verified email, not disabled.\n\n**Errors.** Every error is a problem document (RFC 9457, `application/problem+json`), except the 401 and 403 of an access check, which carry only a `WWW-Authenticate` header. When `type` is present, it links to the documentation of that problem.\n\n**Limits.** Login, sign-up, verification email resend and password reset are rate limited: a 429 gives the seconds to wait in `Retry-After`. A page holds at most 100 items.\n",
    "title" : "Tasks API",
    "version" : "development"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/auth/login" : {
      "post" : {
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponseDto"
                }
              }
            },
            "description" : "Signed in: the token pair."
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/InvalidCredentials"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "security" : [ ],
        "summary" : "Sign in",
        "tags" : [ "Authentication" ]
      }
    },
    "/api/v1/auth/logout" : {
      "post" : {
        "operationId" : "logout",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshTokenRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "security" : [ ],
        "summary" : "Sign out: revoke the refresh token and its family",
        "tags" : [ "Authentication" ]
      }
    },
    "/api/v1/auth/password-reset/confirm" : {
      "post" : {
        "operationId" : "confirmPasswordReset",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordResetConfirmDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "examples" : {
                  "invalid-token" : {
                    "summary" : "Invalid or expired token",
                    "value" : {
                      "status" : 400,
                      "title" : "Invalid or expired token",
                      "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#invalid-token"
                    }
                  },
                  "validation-error" : {
                    "summary" : "Invalid request",
                    "value" : {
                      "errors" : [ {
                        "detail" : "must not be blank",
                        "pointer" : "#/title"
                      } ],
                      "status" : 400,
                      "title" : "Invalid request",
                      "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#validation-error"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description" : "Invalid request: see errors for each invalid value, or type for another problem."
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "security" : [ ],
        "summary" : "Set a new password with the token of the reset email",
        "tags" : [ "Authentication" ]
      }
    },
    "/api/v1/auth/password-reset/request" : {
      "post" : {
        "operationId" : "requestPasswordReset",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordResetRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Accepted"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "security" : [ ],
        "summary" : "Ask for a password reset email; always accepted, whether the account exists or not",
        "tags" : [ "Authentication" ]
      }
    },
    "/api/v1/auth/refresh" : {
      "post" : {
        "operationId" : "refreshTokens",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshTokenRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponseDto"
                }
              }
            },
            "description" : "The new token pair; the refresh token sent is revoked."
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/InvalidCredentials"
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "security" : [ ],
        "summary" : "Trade a refresh token for a new token pair",
        "tags" : [ "Authentication" ]
      }
    },
    "/api/v1/auth/verify-email" : {
      "post" : {
        "operationId" : "verifyEmail",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VerifyEmailRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "examples" : {
                  "invalid-token" : {
                    "summary" : "Invalid or expired token",
                    "value" : {
                      "status" : 400,
                      "title" : "Invalid or expired token",
                      "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#invalid-token"
                    }
                  },
                  "validation-error" : {
                    "summary" : "Invalid request",
                    "value" : {
                      "errors" : [ {
                        "detail" : "must not be blank",
                        "pointer" : "#/title"
                      } ],
                      "status" : 400,
                      "title" : "Invalid request",
                      "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#validation-error"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description" : "Invalid request: see errors for each invalid value, or type for another problem."
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "security" : [ ],
        "summary" : "Verify an email with the token of its link",
        "tags" : [ "Authentication" ]
      }
    },
    "/api/v1/auth/verify-email/resend" : {
      "post" : {
        "operationId" : "resendVerificationEmail",
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "examples" : {
                  "email-already-verified" : {
                    "summary" : "Email already verified",
                    "value" : {
                      "status" : 409,
                      "title" : "Email already verified",
                      "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#email-already-verified"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict: Email already verified."
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "summary" : "Send the verification email again",
        "tags" : [ "Authentication" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "url" : "/"
  } ],
...
}